Kelly Criterion Bet Size

This function calculates the Kelly Criterion and returns the dollar amount to bet in order to maximize returns.

kelly_bet(unit_size, win_prob, odds, type = "dec", kelly_type = "full")

Arguments

unit_size

Unit size of your bankroll, typically 1% of bankroll (100)

win_prob

Probability of winning a bet (0.55)

odds

Odds for a bet (-110)

type

Type of odds. Possible values are:

us, American Odds

dec, Decimal Odds

frac, Fractional Odds

kelly_type

Optional input altering the Kelly formula. Possible values are:

half, Half Kelly

quarter, Quarter Kelly

eighth, Eighth Kelly

Value

Percentage of bankroll to risk on bet

References

https://en.wikipedia.org/wiki/Kelly_criterion

Examples

kelly_bet(unit_size = 100, win_prob = 0.58, odds = -132, type = "us")
#> [1] 256
kelly_bet(unit_size = 50, win_prob = 0.53, odds = -105, type = "us", kelly_type = "Half")
#> [1] 182
kelly_bet(unit_size = 500, win_prob = 0.545, odds = 2.1, type = "dec")
#> [1] 6570
kelly_bet(unit_size = 300, win_prob = 0.27, odds = 5.5, type = "dec", kelly_type = "Quarter")
#> [1] 3234
kelly_bet(unit_size = 250, win_prob = 0.10, odds = 40/1, type = "frac")
#> [1] 1938